# speedup annotation