--- a/tests/test_metadata.py
+++ b/tests/test_metadata.py
@@ -188,10 +188,10 @@ class TestDumpLoad(tests.PyslimTestCase):
 
     def test_load_tables(self):
         for ts in self.get_slim_examples():
-            self.assertTrue(type(ts) is pyslim.SlimTreeSequence)
+            self.assertTrue(isinstance(ts, pyslim.SlimTreeSequence))
             tables = ts.tables
             new_ts = pyslim.load_tables(tables)
-            self.assertTrue(type(new_ts) is pyslim.SlimTreeSequence)
+            self.assertTrue(isinstance(new_ts, pyslim.SlimTreeSequence))
             new_tables = new_ts.tables
             self.assertEqual(tables, new_tables)
 
@@ -200,7 +200,7 @@ class TestDumpLoad(tests.PyslimTestCase):
             fn = ex['basename'] + ".trees"
             # load in msprime then switch
             msp_ts = tskit.load(fn)
-            self.assertTrue(type(msp_ts) is msprime.TreeSequence)
+            self.assertTrue(isinstance(msp_ts, tskit.TreeSequence))
             # transfer tables
             msp_tables = msp_ts.tables
             new_ts = pyslim.load_tables(msp_tables)
@@ -210,13 +210,13 @@ class TestDumpLoad(tests.PyslimTestCase):
             self.assertTableCollectionsEqual(msp_tables, new_tables)
             # convert directly
             new_ts = pyslim.SlimTreeSequence(msp_ts)
-            self.assertTrue(type(new_ts) is pyslim.SlimTreeSequence)
+            self.assertTrue(isinstance(new_ts, pyslim.SlimTreeSequence))
             self.verify_times(msp_ts, new_ts)
             new_tables = new_ts.tables
             self.assertTableCollectionsEqual(msp_tables, new_tables)
             # load to pyslim from file
             slim_ts = pyslim.load(fn)
-            self.assertTrue(type(slim_ts) is pyslim.SlimTreeSequence)
+            self.assertTrue(isinstance(slim_ts, pyslim.SlimTreeSequence))
             slim_tables = slim_ts.tables
             self.assertTableCollectionsEqual(msp_tables, slim_tables)
             self.assertEqual(slim_ts.slim_generation, new_ts.slim_generation)