It was so TERRIBLY SLOW, turns out because of all those python for loops. This went form
```
$ tests/benchmark/annotation.sh 
setup:
1 loop, best of 5: 762 msec per loop
tskit:
2 loops, best of 5: 116 msec per loop
pyslim:
1 loop, best of 5: 25.6 sec per loop
```
to
```
$ tests/benchmark/annotation.sh 
setup:
1 loop, best of 5: 735 msec per loop
tskit:
2 loops, best of 5: 125 msec per loop
pyslim:
2 loops, best of 5: 115 msec per loop
```